Revise process chaining clarity.
The process method for process chaining is more clearly written.
However the revision to this has removed the ready flag was removed
from poller runners. This means stacking and overlap may occur. It
is not confirmed as to how unstable this makes the overall aggregation
routine.

A runner id is now base64 encoded so the is less complex.